+++ lib/doorkeeper/config.rb
@@ -337,7 +337,8 @@ module Doorkeeper
     #
     option :realm,                          default: "Doorkeeper"
 
-    # Issuer URL for the OAuth Authorization Server
+    # Issuer URL advertised in the OAuth 2.0 Authorization Server Metadata
+    # (RFC 8414). When nil, the request base URL is used instead.
     option :issuer,                         default: nil
 
     # Forces the usage of the HTTPS protocol in non-native redirect uris
@@ -416,9 +417,9 @@ module Doorkeeper
     option :application_class,
            default: "Doorkeeper::Application"
 
-    # Allows setting a hash of custom data on the OAuth 2.0 Authorization
-    # Server Metadata discovery response.
-    option :custom_discovery_data, default: {}
+    # Allows setting a hash of custom data merged into the OAuth 2.0
+    # Authorization Server Metadata response (RFC 8414).
+    option :custom_metadata, default: {}
 
     # Allows to set blank redirect URIs for Applications in case
     # server configured to use URI-less grant flows.