+++ lib/doorkeeper/rails/routes/mapping.rb
@@ -11,7 +11,7 @@ module Doorkeeper
             authorizations: "doorkeeper/authorizations",
             applications: "doorkeeper/applications",
             authorized_applications: "doorkeeper/authorized_applications",
-            discovery: "doorkeeper/discovery",
+            metadata: "doorkeeper/metadata",
             tokens: "doorkeeper/tokens",
             token_info: "doorkeeper/token_info",
           }