+++ spec/dummy/app/controllers/custom_discovery_controller.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class CustomDiscoveryController < ::ApplicationController
-  def show
-    render nothing: true
-  end
-end