+++ spec/dummy/app/controllers/custom_metadata_controller.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class CustomMetadataController < ::ApplicationController
+  def show
+    head :ok
+  end
+end