+++ spec/routing/custom_controller_routes_spec.rb
@@ -17,7 +17,7 @@ RSpec.describe "Custom controller for routes" do
                       tokens: "custom_authorizations",
                       applications: "custom_authorizations",
                       token_info: "custom_authorizations",
-                      discovery: "custom_discovery"
+                      metadata: "custom_metadata"
 
           as authorizations: "custom_auth",
              tokens: "custom_token",
@@ -31,7 +31,7 @@ RSpec.describe "Custom controller for routes" do
                       tokens: "custom_authorizations",
                       applications: "custom_authorizations",
                       token_info: "custom_authorizations",
-                      discovery: "custom_discovery"
+                      metadata: "custom_metadata"
 
           as authorizations: "custom_auth",
              tokens: "custom_token",
@@ -44,7 +44,7 @@ RSpec.describe "Custom controller for routes" do
           controllers authorizations: "custom_authorizations",
                       tokens: "custom_authorizations",
                       token_info: "custom_authorizations",
-                      discovery: "custom_discovery"
+                      metadata: "custom_metadata"
 
           as authorizations: "custom_auth",
              tokens: "custom_token",
@@ -86,10 +86,9 @@ RSpec.describe "Custom controller for routes" do
     expect(get("/inner_space/scope/token/info")).to route_to("custom_authorizations#show")
   end
 
-  it "GET /inner_space/.well-known/oauth-authorization-server route to show Discovery controller" do
-    expect(get("/space/.well-known/oauth-authorization-server")).to route_to("custom_discovery#show")
+  it "GET /inner_space/.well-known/oauth-authorization-server routes to custom Metadata controller" do
+    expect(get("/inner_space/.well-known/oauth-authorization-server")).to route_to("custom_metadata#show")
   end
-
 
   it "GET /space/oauth/authorize routes to custom authorizations controller" do
     expect(get("/space/oauth/authorize")).to route_to("custom_authorizations#new")
@@ -123,8 +122,8 @@ RSpec.describe "Custom controller for routes" do
     expect(get("/space/oauth/token/info")).to route_to("custom_authorizations#show")
   end
 
-  it "GET /space/.well-known/oauth-authorization-server route to show Discovery controller" do
-    expect(get("/space/.well-known/oauth-authorization-server")).to route_to("custom_discovery#show")
+  it "GET /space/.well-known/oauth-authorization-server routes to custom Metadata controller" do
+    expect(get("/space/.well-known/oauth-authorization-server")).to route_to("custom_metadata#show")
   end
 
   it "POST /outer_space/oauth/token is not be routable" do
@@ -143,7 +142,7 @@ RSpec.describe "Custom controller for routes" do
     expect(get("/outer_space/oauth/token/info")).not_to be_routable
   end
 
-  it "GET /outer_space/.well-known/oauth-authorization-server route to show Discovery controller" do
-    expect(get("/outer_space/.well-known/oauth-authorization-server")).to route_to("custom_discovery#show")
+  it "GET /outer_space/.well-known/oauth-authorization-server routes to custom Metadata controller" do
+    expect(get("/outer_space/.well-known/oauth-authorization-server")).to route_to("custom_metadata#show")
   end
 end