+++ spec/routing/scoped_routes_spec.rb
@@ -54,7 +54,7 @@ RSpec.describe "Scoped routes" do
     expect(post("/scope/introspect")).not_to be_routable
   end
 
-  it "GET /.well-known/oauth-authorization-server route to show Discovery controller" do
-    expect(get("/.well-known/oauth-authorization-server")).to route_to("doorkeeper/discovery#show")
+  it "GET /.well-known/oauth-authorization-server routes to Metadata controller" do
+    expect(get("/.well-known/oauth-authorization-server")).to route_to("doorkeeper/metadata#show")
   end
 end